Create mazeProblem
Solving a Depth-first search coding problem that the .... are spots where someone could walk through a maze, and the * and # are barriers. Given a choice of being allowed to go through # or *. After removing the selected barrier you try to go to the end of the maze which is the bottom right corner. This gives the shortest distance to solve the maze depending on which you are allowed to pass through.